+++ idea/testData/codeInsight/overrideImplement/copyKDoc.kt
@@ -1,0 +1,11 @@
+// COPY_DOC
+abstract class A {
+    /**
+     * @see TEST
+     */
+    abstract fun foo()
+}
+
+class B : A() {
+    <caret>
+}