+++ idea/testData/multiFileIntentions/reconcilePackageWithDirectory/addQuotation/after/in/foo/fun/usagesWithImports.java
@@ -1,0 +1,10 @@
+package in.foo.fun;
+
+import static in.foo.fun.TestKt.foo;
+
+class Test {
+    static void test() {
+        new Foo();
+        foo();
+    }
+}