+++ idea/testData/multiFileIntentions/reconcilePackageWithDirectory/addQuotation/before/in/foo/fun/usagesWithFqNames.java
@@ -1,0 +1,8 @@
+package in.foo.fun;
+
+class Test {
+    static void test() {
+        new bar.Foo();
+        bar.TestKt.foo();
+    }
+}