+++ idea/testData/refactoring/move/kotlin/moveDirectoryWithQuotation/before/in/p1/p/K.kt
@@ -1,0 +1,5 @@
+package `in`.p1.p
+
+public class K {
+    val j = J()
+}