+++ idea/testData/refactoring/rename/renameDefaultJavaAnnotationMethod/before/test/KAnn.kt
@@ -1,0 +1,7 @@
+package test
+
+@JAnn("abc")
+fun test1() {}
+
+@JAnn(value = "abc")
+fun test2() {}