+++ idea/testData/refactoring/rename/renameDefaultKotlinAnnotationMethod/after/test/JAnn.java
@@ -1,0 +1,9 @@
+package test;
+
+class Test {
+    @KAnn(valueNew = "abc")
+    void test1() { }
+
+    @KAnn(valueNew = "abc")
+    void test2() { }
+}